+++ src/webrtc/datachannel.ts
@@ -7,39 +7,12 @@
 // platform compatibility library webrtc-adaptor.js (from:
 // https://code.google.com/p/webrtc/source/browse/stable/samples/js/base/adapter.js)
 
+import DataChannels = require('datachannel.i');
 import Handler = require('../handler/queue');
 import ArrayBuffers = require('../arraybuffers/arraybuffers');
 import Logging = require('../logging/logging');
 
-
-export interface DataChannel {
-  // Guarenteed to be invarient for the life of the data channel.
-  getLabel : () => string;
-
-  // Promise for when the data channel has been openned.
-  onceOpened : Promise<void>;
-
-  // Promise for when the data channel has been closed (only fulfilled after
-  // the data channel has been openned).
-  // NOTE: There exists a bug in Chrome prior to version 37 which prevents
-  //       this from fulfilling on the remote peer.
-  onceClosed : Promise<void>;
-
-  // Data from the peer. No data will be added to the queue after |onceClosed|
-  // is fulfilled.
-  dataFromPeerQueue :Handler.QueueHandler<Data, void>;
-
-  // Send data; promise returns when all the data has been passed on to the
-  // undertlying network layer for ending.
-  send(data:Data) : Promise<void>;
-
-  // Closes this data channel.
-  // A channel cannot be re-opened once this has been called.
-  close() : void;
-
-  toString() : string;
-}
-
+import Data = DataChannels.Data;
 
 // Messages are limited to a 16KB length by SCTP; we use 15k for safety.
 // TODO: test if we can up this to 16k; test the edge-cases!
@@ -57,18 +30,6 @@ var PC_QUEUE_LIMIT = 1024 * 250;
 // don't support trying to send array's bigger than that.
 var MAX_MESSAGE_SIZE = Math.pow(2, 53);
 
-// Data sent to or received from a peer on a data channel in the peer
-// connection.
-export interface Data {
-  str ?:string;
-  buffer ?:ArrayBuffer;
-  // TODO: add when supported by WebRtc in Chrome and FF.
-  // https://code.google.com/p/webrtc/issues/detail?id=2276
-  //
-  // bufferView  ?:ArrayBufferView;
-  // blob        ?:Blob
-  // domString   ?:DOMString
-}
 interface StringData { str :string; }
 interface BufferData { buffer :ArrayBuffer; }
 
@@ -77,7 +38,17 @@ var log :Logging.Log = new Logging.Log('DataChannel');
 // Wrapper for a WebRtc Data Channels:
 // http://dev.w3.org/2011/webrtc/editor/webrtc.html#rtcdatachannel
 //
-export class DataChannelClass implements DataChannel {
+class DataChannelClass implements DataChannels.Channel {
+
+  // Static, async constructor that returns a DataChannel ready for use.
+  // |id| is the GUID generated by Freedom to identify the underlying
+  //     browser object.
+  public static createFromFreedomId(id:string) : Promise<DataChannels.Channel> {
+    var rtcDataChannel = freedom['core.rtcdatachannel'](id);
+    return rtcDataChannel.getLabel().then((label:string) => {
+      return new DataChannelClass(rtcDataChannel, label, id);
+    });
+  }
 
   public dataFromPeerQueue      :Handler.Queue<Data,void>;
 
@@ -266,13 +237,4 @@ export class DataChannelClass implements DataChannel {
   }
 }  // class DataChannelClass
 
-
-// Static, async constructor that returns a DataChannel ready for use.
-// |id| is the GUID generated by Freedom to identify the underlying
-//     browser object.
-export function fromId(id:string) : Promise<DataChannel> {
-  var rtcDataChannel = freedom['core.rtcdatachannel'](id);
-  return rtcDataChannel.getLabel().then((label:string) => {
-    return new DataChannelClass(rtcDataChannel, label, id);
-  });
-}
+export = DataChannelClass;
