+++ src/webrtc/webrtc.ts
@@ -1,0 +1,9 @@
+import PeerConnectionInterfaces = require('peerconnection.i');
+import PeerConnectionClass = require('peerconnection');
+
+export function createPeerConnection(
+      config:PeerConnectionInterfaces.PeerConnectionConfig)
+    : PeerConnectionInterfaces.PeerConnection<
+        PeerConnectionInterfaces.SignallingMessage> {
+  return new PeerConnectionClass(config);
+}
